Get, Update, or Delete an Earth Engine Asset by ID

Get, Update, or Delete an Earth Engine Asset by ID

gd_get_asset(x, silent = FALSE)

gd_update_asset(
  x,
  asset,
  update = c("start_time", "end_time", "properties"),
  silent = FALSE
)

gd_delete_asset(x, silent = FALSE)

Arguments

x

Asset ID name

silent

Silence errors? Default: FALSE

asset

Used only for gd_update_asset(): a named list, with names representing elements of x to replace. Only "start_time", "end_time", and "properties" fields can be updated.

update

Used only for gd_update_asset(): A character vector of field names to update. Default: "start_time", and "end_time" to update timestamps; and "properties" to update all properties.

Value

try-error on error. gd_get_asset(): a named list containing information and properties of an Earth Engine asset

gd_update_asset(): This function is called for side-effects (updates the specified asset fields)

gd_delete_asset(): This function is called for side-effects (deletes the specified asset)

Examples

if (FALSE) { # \dontrun{
# get asset from project by ID
a <- gd_get_asset("projects/your-project-name/assets/YOUR_ASSET_ID")
} # }
if (FALSE) { # \dontrun{
# change description in `"properties"`
a$properties$description <- "foo"

# update asset
gd_update_asset("projects/your-project-name/assets/YOUR_ASSET_ID", a, "properties")
} # }
if (FALSE) { # \dontrun{
# remove an asset from project
gd_delete_asset("projects/your-project-name/assets/YOUR_ASSET_ID")
} # }