Export image to Earth Engine Asset, Google Cloud Storage Bucket, or Google Drive

Exports an encapsulated image to the destination specified by type, folder and filename

gd_export(
  x,
  filename,
  type = "drive",
  folder = dirname(filename),
  region,
  wait = TRUE,
  overwrite = TRUE,
  ...
)

Arguments

x

An object that inherits from geedim.image.ImageAccessor (for geedim >= 2.0.0) or geedim.download.BaseImage (for geedim < 2.0.0). See \link{geedim-versions} for more details.

filename

Output filename. If type is "asset" and folder is not specified, filename should be a valid Earth Engine asset ID.

type

Export type. Defaults to "drive"; other options include "asset", and "cloud". See gd_export_types()

folder

Destination folder. Defaults to dirname(filename).

region

Region e.g. from gd_bbox() or gd_region()

wait

Wait for completion? Default: TRUE

overwrite

Overwrite existing? Default: TRUE. Only supported for geedim >= 2.0.0.

...

Additional arguments to geedim.image.ImageAccessor.toGoogleCloud() (geedim >= 2.0.0) or geedim.download.BaseImage.export() (geedim 1.x.x)

Value

an ee.batch.Task object

Details

See the geedim.image.ImageAccessor.toGoogleCloud() documentation for details on additional arguments. Requires 'geedim' >1.6.0.

Examples

if (FALSE) { # \dontrun{
if (gd_is_initialized()) {
 r <- gd_bbox(
   xmin = -120.6032,
   xmax = -120.5377,
   ymin = 38.0807,
   ymax = 38.1043
 )

 i <- gd_image_from_id('CSP/ERGo/1_0/US/CHILI')

 ## export to Google Drive (default `type="drive"`)
 # res <- gd_export(i, filename = "RGEEDIM_TEST", scale = 100, region = r)

 ## export to `type="asset"`, then download by ID (stored in project assets)
 # res <- gd_export(
 #   i,
 #   "RGEEDIM_TEST",
 #   type = "asset",
 #   folder = "rgeedim-demo",
 #   scale = 100,
 #   region = r
 # )
 # gd_download("projects/rgeedim-demo/assets/RGEEDIM_TEST", filename = "test.tif")

 ## export to Google Cloud Bucket with `type="cloud"`,
 ##   where `folder` is the bucket path without `"gs://"`
 # res <- gd_export(i, filename = "RGEEDIM_TEST.tif", type = "cloud",
 #                  folder = "your-bucket-name", scale = 100, region = r)
}
} # }