Initialize geedim

Calls geedim Initialize() method. This method should be called at the beginning of each session.

gd_initialize(
  private_key_file = NULL,
  credentials = "persistent",
  cloud_api_key = NULL,
  url = "https://earthengine-highvolume.googleapis.com",
  opt_url = NULL,
  http_transport = NULL,
  project = NULL,
  quiet = TRUE
)

gd_is_initialized()

Arguments

private_key_file

character. Optional: Path to JSON file containing client information and private key. Alternately, the contents of a JSON file. Instead of setting this argument you may specify EE_SERVICE_ACC_PRIVATE_KEY environment variable with path to JSON file.

credentials

Default: 'persistent' uses credentials already stored in the filesystem, or raise an explanatory exception guiding the user to create those credentials.

cloud_api_key

An optional API key to use the Cloud API. Default: NULL.

url

The base url for the EarthEngine REST API to connect to. Defaults to "High Volume" endpoint: "https://earthengine-highvolume.googleapis.com"

opt_url

(deprecated) Use url.

http_transport

The HTTP transport method to use when making requests. Default: NULL

project

The client project ID or number to use when making API calls. Default: NULL

quiet

Suppress error messages on load? Default: FALSE

Value

gd_initialize(): try-error (invisibly) on error.

gd_is_initialized(): logical. TRUE if initialized successfully.

See also

Examples

if (FALSE) {
gd_initialize()
}
gd_is_initialized()
#> [1] FALSE